+++ NEWS.md
@@ -11,6 +11,7 @@
 * [SIP2-291](https://folio-org.atlassian.net/browse/SIP2-291): Implement a single login operation per session
 * [SIP2-296](https://folio-org.atlassian.net/browse/SIP2-296): Update hard-coded currency list
 * [SIP2-242](https://folio-org.atlassian.net/browse/SIP2-242): Skip null barcodes in PatronInformation response
+* [SIP2-294](https://folio-org.atlassian.net/browse/SIP2-294): Replace com.github.mikelee2082:vertx-config-s3 with folio-s3-client configuration
 
 
 ## 3.4.0 2025-03-14